A clunky, but working, distortion test.
The test currently validates that a 0-distortion camera results in a calibration
that is also (very close to) 0.

The poses used to calibrate need to be refined and the code is currently a mess
