+++ gazebo_plugins/test/camera/distortion.test
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <launch>
-  <arg name="gui" default="false" />
+  <arg name="gui" default="true" />
 
   <param name="/use_sim_time" value="true" />
 
@@ -13,7 +13,7 @@
   </group>
 
   <test test-name="distortion_test" pkg="gazebo_plugins" type="distortion_test"
-      clear_params="true" time-limit="30.0">
+      clear_params="true" time-limit="60.0">
     <param name="calibration_settings_file" value="$(find gazebo_plugins)/test/camera/distortion_calibration_params.xml" />
   </test>
 </launch>
